Admin Panel/requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.12.2
cryptography==2.6.1
Django==2.1.7
djangorestframework==3.9.2
future==0.17.1
pycparser==2.19
python-telegram-bot==11.1.0
pytz==2018.9
six==1.12.0
